requirements.txt
agno==1.1.12
annotated-types==0.7.0
anthropic==0.49.0
anyio==4.8.0
beautifulsoup4==4.13.3
boto3==1.37.8
botocore==1.37.8
Brotli==1.1.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
docstring_parser==0.16
duckduckgo_search==7.5.1
email_validator==2.2.0
fastapi==0.115.8
fastapi-cli==0.0.7
frozendict==2.4.6
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
html5lib==1.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
lancedb==0.18.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
multitasking==0.0.11
numpy==2.2.2
openai==1.66.3
overrides==7.7.0
packaging==24.2
pandas==2.2.3
peewee==3.17.8
pgvector==0.3.6
phidata==2.7.10
platformdirs==4.3.6
primp==0.14.0
psycopg==3.2.5
psycopg-binary==3.2.5
psycopg2-binary==2.9.10
pyarrow==19.0.0
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
Pygments==2.19.1
pylance==0.22.0
pypdf==5.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.20
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rich-toolkit==0.13.2
s3transfer==0.11.4
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
socksio==1.0.0
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.45.3
tantivy==0.22.0
tavily-python==0.5.1
tiktoken==0.9.0
tomli==2.2.1
tqdm==4.67.1
typer==0.15.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
watchfiles==1.0.4
webencodings==0.5.1
websockets==14.2
yfinance==0.2.52
